+++ @generated/graphql/index.ts
@@ -2372,23 +2372,34 @@ export type StoreCurrency = {
   symbol: Scalars['String']
 }
 
-/** Search facet information. */
-export type StoreFacet = {
+export type StoreFacet = StoreFacetBoolean | StoreFacetRange
+
+/** Search facet boolean information. */
+export type StoreFacetBoolean = {
   /** Facet key. */
   key: Scalars['String']
   /** Facet label. */
   label: Scalars['String']
-  /** Facet type. Possible values are `BOOLEAN` and `RANGE`. */
-  type: StoreFacetType
   /** Array with information on each facet value. */
-  values: Array<StoreFacetValue>
+  values: Array<StoreFacetValueBoolean>
+}
+
+/** Search facet range information. */
+export type StoreFacetRange = {
+  /** Facet key. */
+  key: Scalars['String']
+  /** Facet label. */
+  label: Scalars['String']
+  max: StoreFacetValueRange
+  /** Array with information on each facet value. */
+  min: StoreFacetValueRange
 }
 
 /** Search facet type. */
 export type StoreFacetType = 'BOOLEAN' | 'RANGE'
 
 /** Information of a specific facet value. */
-export type StoreFacetValue = {
+export type StoreFacetValueBoolean = {
   /** Facet value label. */
   label: Scalars['String']
   /** Number of items with this facet. */
@@ -2397,6 +2408,11 @@ export type StoreFacetValue = {
   selected: Scalars['Boolean']
   /** Facet value. */
   value: Scalars['String']
+}
+
+export type StoreFacetValueRange = {
+  absolute: Scalars['Float']
+  selected: Scalars['Float']
 }
 
 /** Image. */
@@ -2669,10 +2685,10 @@ export type ProductSummary_ProductFragment = {
   }
 }
 
-export type Filter_FacetsFragment = {
+export type Filter_Facets_StoreFacetBoolean_Fragment = {
+  __typename: 'StoreFacetBoolean'
   key: string
   label: string
-  type: StoreFacetType
   values: Array<{
     label: string
     value: string
@@ -2680,6 +2696,18 @@ export type Filter_FacetsFragment = {
     quantity: number
   }>
 }
+
+export type Filter_Facets_StoreFacetRange_Fragment = {
+  __typename: 'StoreFacetRange'
+  key: string
+  label: string
+  min: { selected: number; absolute: number }
+  max: { selected: number; absolute: number }
+}
+
+export type Filter_FacetsFragment =
+  | Filter_Facets_StoreFacetBoolean_Fragment
+  | Filter_Facets_StoreFacetRange_Fragment
 
 export type ProductDetailsFragment_ProductFragment = {
   sku: string
@@ -2720,42 +2748,27 @@ export type ProductGalleryQueryQueryVariables = Exact<{
 
 export type ProductGalleryQueryQuery = {
   search: {
-    products: {
-      pageInfo: { totalCount: number }
-      edges: Array<{
-        node: {
-          slug: string
-          sku: string
-          name: string
-          gtin: string
-          id: string
-          brand: { name: string; brandName: string }
-          isVariantOf: { productGroupID: string; name: string }
-          image: Array<{ url: string; alternateName: string }>
-          offers: {
-            lowPrice: number
-            offers: Array<{
-              availability: string
-              price: number
-              listPrice: number
-              quantity: number
-              seller: { identifier: string }
-            }>
-          }
+    products: { pageInfo: { totalCount: number } }
+    facets: Array<
+      | {
+          __typename: 'StoreFacetBoolean'
+          key: string
+          label: string
+          values: Array<{
+            label: string
+            value: string
+            selected: boolean
+            quantity: number
+          }>
         }
-      }>
-    }
-    facets: Array<{
-      key: string
-      label: string
-      type: StoreFacetType
-      values: Array<{
-        label: string
-        value: string
-        selected: boolean
-        quantity: number
-      }>
-    }>
+      | {
+          __typename: 'StoreFacetRange'
+          key: string
+          label: string
+          min: { selected: number; absolute: number }
+          max: { selected: number; absolute: number }
+        }
+    >
   }
 }